Moved ColumnConstraintInformation to create_statement.py
